Rename _marks/scatter to _marks/dot
